Revert "feat: save using shortcuts"
This reverts commit d4e2b5baa92d8f2046ad31250c46386a96a7ec46.
