+++ src/app/utils.ts
@@ -1,6 +1,4 @@
-import type { AppState } from './context/appState';
-
-function download(blob: Blob, name: string) {
+export function download(blob: Blob, name: string) {
   const link = document.createElement('a');
 
   const url = URL.createObjectURL(blob);
@@ -12,18 +10,4 @@ function download(blob: Blob, name: string) {
   link.click();
   document.body.removeChild(link);
   URL.revokeObjectURL(url);
-}
-export function saveHandler(
-  appState: AppState,
-  filename = 'file',
-  spaceIndent = 0,
-) {
-  const data = JSON.stringify(
-    { data: appState.data, view: appState.view },
-    (_key, value) =>
-      ArrayBuffer.isView(value) ? Array.from(value as any) : value,
-    spaceIndent,
-  );
-  const blob = new Blob([data], { type: 'text/plain' });
-  download(blob, `${filename}.ium`);
 }
